requirements/base.txt
PyJWT>=1.5.2,<2.0.0
